+++ src/main/kotlin/racing/model/move/MoveThresholdStrategy.kt
@@ -5,12 +5,16 @@ import racing.model.generator.Generator
 class MoveThresholdStrategy(
     private val generator: Generator,
 ) : MoveStrategy {
-    private val threshold = 4
+
     override fun movable(): Boolean {
         return isOverThreshold(generator.generate())
     }
 
     private fun isOverThreshold(value: Int): Boolean {
-        return value >= threshold
+        return value >= THRESHOLD
+    }
+
+    companion object {
+        private const val THRESHOLD = 4
     }
 }
